requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
dataclasses-json==0.6.4
decorator==4.4.2
deepgram-sdk==3.7.3
deprecation==2.1.0
distro==1.9.0
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-speech==2.24.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.3
httplib2==0.22.0
httpx==0.26.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
lxml==5.2.2
marshmallow==3.20.2
moviepy==1.0.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.12.0
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
praw==7.8.0
prawcore==2.4.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.2
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
pyparsing==3.1.1
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
selectolax==0.3.20
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.27
srt==3.5.3
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.9.0
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.0
vosk==0.3.44
webdriver-manager==4.0.1
websocket-client==1.7.0
websockets==13.1
wsproto==1.2.0
yarl==1.9.4
